[refactor] Episode 도메인 패키지 구조 개선 - #9
