--- novelservice/src/main/java/com/iucyh/novelservice/common/exception/episode/EpisodeNotFound.java
@@ -1,31 +1,0 @@
-package com.iucyh.novelservice.common.exception.episode;
-
-import com.iucyh.novelservice.common.exception.ServiceException;
-
-import java.util.Map;
-
-public class EpisodeNotFound extends ServiceException {
-
-    public EpisodeNotFound(long episodeId) {
-        super(
-                EpisodeErrorCode.EPISODE_NOT_FOUND,
-                Map.of("episodeId", episodeId)
-        );
-    }
-
-    private <T> EpisodeNotFound(String message, String fieldName, T value) {
-        super(
-                EpisodeErrorCode.EPISODE_NOT_FOUND,
-                message,
-                Map.of(fieldName, value)
-        );
-    }
-
-    public static EpisodeNotFound withEpisodeNumber(int episodeNumber) {
-        return new EpisodeNotFound(
-                "Episode not found with this episode number",
-                "episodeNumber",
-                episodeNumber
-        );
-    }
-}
