--- novelservice/src/main/java/com/iucyh/novelservice/dto/episode/request/CreateEpisodeRequest.java
@@ -1,24 +1,0 @@
-package com.iucyh.novelservice.dto.episode.request;
-
-import com.iucyh.novelservice.common.constant.EpisodeConstants;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
-import jakarta.validation.constraints.Size;
-
-import static com.iucyh.novelservice.common.constant.EpisodeConstants.*;
-
-public record CreateEpisodeRequest(
-
-        @NotBlank
-        @Size(min = EPISODE_TITLE_LENGTH_MIN, max = EpisodeConstants.EPISODE_TITLE_LENGTH_MAX)
-        String title,
-
-        @NotNull
-        @Size(max = EPISODE_DESC_LENGTH_MAX)
-        String description,
-
-        @NotBlank
-        @Size(min = EPISODE_CONTENT_LENGTH_MIN, max = EPISODE_CONTENT_LENGTH_MAX)
-        String content
-) {}
