--- novelservice/src/main/java/com/iucyh/novelservice/dto/episode/request/EpisodePagingRequest.java
@@ -1,21 +1,0 @@
-package com.iucyh.novelservice.dto.episode.request;
-
-import jakarta.validation.constraints.Positive;
-import org.hibernate.validator.constraints.Range;
-import org.springframework.web.bind.annotation.RequestParam;
-
-public record EpisodePagingRequest(
-
-        @RequestParam("last-episode")
-        @Positive
-        Integer lastEpisode,
-
-        @Range(min = 1, max = 100, message = "Limit must be between 1 and 100")
-        Integer limit
-) {
-    public EpisodePagingRequest {
-        if (limit == null) {
-            limit = 50;
-        }
-    }
-}
