--- novelservice/src/main/java/com/iucyh/novelservice/dto/episode/response/EpisodeDetailResponse.java
@@ -1,3 +1,0 @@
-package com.iucyh.novelservice.dto.episode.response;
-
-public record EpisodeDetailResponse(String content) {}
