+++ novelservice/src/main/java/com/iucyh/novelservice/episode/constant/EpisodeConstants.java
@@ -1,0 +1,12 @@
+package com.iucyh.novelservice.episode.constant;
+
+public abstract class EpisodeConstants {
+
+    public static final int EPISODE_TITLE_LENGTH_MIN = 1;
+    public static final int EPISODE_TITLE_LENGTH_MAX = 32;
+
+    public static final int EPISODE_DESC_LENGTH_MAX = 20;
+
+    public static final int EPISODE_CONTENT_LENGTH_MIN = 100;
+    public static final int EPISODE_CONTENT_LENGTH_MAX = 20000;
+}
