+++ novelservice/src/main/java/com/iucyh/novelservice/episode/repository/projection/EpisodeDetail.java
@@ -1,0 +1,7 @@
+package com.iucyh.novelservice.episode.repository.projection;
+
+public interface EpisodeDetail {
+
+    Long getId();
+    String getContent();
+}
