+++ novelservice/src/main/java/com/iucyh/novelservice/episode/web/dto/request/UpdateEpisodeRequest.java
@@ -1,0 +1,16 @@
+package com.iucyh.novelservice.episode.web.dto.request;
+
+import com.iucyh.novelservice.common.validator.notblank.NotBlankIfPresent;
+import jakarta.validation.constraints.Size;
+
+import static com.iucyh.novelservice.episode.constant.EpisodeConstants.*;
+
+public record UpdateEpisodeRequest(
+
+        @NotBlankIfPresent
+        @Size(min = EPISODE_TITLE_LENGTH_MIN, max = EPISODE_TITLE_LENGTH_MAX)
+        String title,
+
+        @Size(max = EPISODE_DESC_LENGTH_MAX)
+        String description
+) {}
