+++ novelservice/src/main/java/com/iucyh/novelservice/episode/web/dto/response/EpisodeDetailResponse.java
@@ -1,0 +1,3 @@
+package com.iucyh.novelservice.episode.web.dto.response;
+
+public record EpisodeDetailResponse(String content) {}
