+++ novelservice/src/main/java/com/iucyh/novelservice/novel/web/controller/NovelController.java
@@ -3,19 +3,19 @@ package com.iucyh.novelservice.novel.web.controller;
 import com.iucyh.novelservice.novel.enumtype.NovelCategory;
 import com.iucyh.novelservice.dto.PagingResultDto;
 import com.iucyh.novelservice.dto.SuccessDto;
-import com.iucyh.novelservice.dto.episode.request.CreateEpisodeRequest;
-import com.iucyh.novelservice.dto.episode.request.EpisodePagingRequest;
-import com.iucyh.novelservice.dto.episode.request.UpdateEpisodeDetailRequest;
-import com.iucyh.novelservice.dto.episode.request.UpdateEpisodeRequest;
-import com.iucyh.novelservice.dto.episode.response.EpisodeDetailResponse;
-import com.iucyh.novelservice.dto.episode.response.EpisodeResponse;
+import com.iucyh.novelservice.episode.web.dto.request.CreateEpisodeRequest;
+import com.iucyh.novelservice.episode.web.dto.request.EpisodePagingRequest;
+import com.iucyh.novelservice.episode.web.dto.request.UpdateEpisodeDetailRequest;
+import com.iucyh.novelservice.episode.web.dto.request.UpdateEpisodeRequest;
+import com.iucyh.novelservice.episode.web.dto.response.EpisodeDetailResponse;
+import com.iucyh.novelservice.episode.web.dto.response.EpisodeResponse;
 import com.iucyh.novelservice.novel.web.dto.request.CreateNovelRequest;
 import com.iucyh.novelservice.novel.web.dto.request.NovelPagingRequest;
 import com.iucyh.novelservice.novel.web.dto.response.NovelLikeCountResponse;
 import com.iucyh.novelservice.novel.web.dto.request.UpdateNovelRequest;
 import com.iucyh.novelservice.novel.web.dto.response.NovelResponse;
-import com.iucyh.novelservice.service.episode.EpisodeQueryService;
-import com.iucyh.novelservice.service.episode.EpisodeService;
+import com.iucyh.novelservice.episode.service.EpisodeQueryService;
+import com.iucyh.novelservice.episode.service.EpisodeService;
 import com.iucyh.novelservice.novel.service.NovelQueryService;
 import com.iucyh.novelservice.novel.service.NovelService;
 import jakarta.validation.Valid;
