--- novelservice/src/main/java/com/iucyh/novelservice/repository/episode/projection/EpisodeDetail.java
@@ -1,7 +1,0 @@
-package com.iucyh.novelservice.repository.episode.projection;
-
-public interface EpisodeDetail {
-
-    Long getId();
-    String getContent();
-}
